Move all implementation to its own files
- Extract HTTPClientError enum to NetworkingInterface/HTTPClientError.swift
- Extract HTTPRequest protocol to NetworkingInterface/HTTPRequest.swift
- Extract HTTPMethod enum to NetworkingInterface/HTTPMethod.swift
- Extract HTTPBody struct to NetworkingInterface/HTTPBody.swift
- Extract HTTPResult struct to NetworkingInterface/HTTPResult.swift
- Extract HTTPClient protocol to NetworkingInterface/HTTPClient.swift
- Extract HTTPErrorMapper to Networking/HTTPErrorMapper.swift
- Extract URLSessionHttpClient to Networking/URLSessionHttpClient.swift
- Remove generic Module.swift and ModuleInterface.swift files
- Update tests to import from proper modules
