--- Modules/Networking/Networking/Module.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public final class Module {
-    public init() {}
-}
