+++ Modules/Networking/NetworkingInterface/HTTPMethod.swift
@@ -1,0 +1,16 @@
+//
+//  HTTPMethod.swift
+//  NetworkingInterface
+//
+//  Created by MacPro on 21/06/25.
+//  Copyright © 2025 Cesar Lima Consulting. All rights reserved.
+//
+
+import Foundation
+
+public enum HTTPMethod: String {
+    case get = "GET"
+    case post = "POST"
+    case put = "PUT"
+    case delete = "DELETE"
+}
