+++ Modules/Networking/NetworkingInterface/HTTPRequest.swift
@@ -1,0 +1,31 @@
+//
+//  HTTPRequest.swift
+//  NetworkingInterface
+//
+//  Created by MacPro on 21/06/25.
+//  Copyright © 2025 Cesar Lima Consulting. All rights reserved.
+//
+
+import Foundation
+
+public protocol HTTPRequest {
+    var host: String { get }
+    var path: String { get }
+    var method: HTTPMethod { get }
+    var headers: [String: String] { get }
+    var body: HTTPBody { get }
+}
+
+extension HTTPRequest {
+    public func asURLRequest() throws -> URLRequest {
+        var components = URLComponents()
+        components.scheme = "https"
+        components.host = host
+        components.path = path
+        var request = URLRequest(url: components.url!)
+        request.httpMethod = method.rawValue
+        request.allHTTPHeaderFields = headers
+        request.httpBody = try body.encode()
+        return request
+    }
+}
