--- Modules/Networking/NetworkingInterface/ModuleInterface.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public protocol ModuleInterface {
-    func start()
-}
