+++ Modules/Networking/NetworkingTests/URLSessionHttpClientTests.swift
@@ -1,164 +1,7 @@
 import XCTest
 import Foundation
+import NetworkingInterface
 @testable import Networking
-
-final class URLSessionHttpClient {
-    private let session: URLSession
-
-    init(session: URLSession = .shared) {
-        self.session = session
-    }
-
-    func perform(_ request: HTTPRequest) async throws -> HTTPResult {
-        do {
-            let (data, response) = try await session.data(for: request.asURLRequest())
-            
-            guard let httpResponse = response as? HTTPURLResponse else {
-                throw URLError(.badServerResponse)
-            }
-            
-            if !(200...299).contains(httpResponse.statusCode) {
-                throw HTTPErrorMapper.mapHTTPResponse(httpResponse)
-            }
-            
-            return HTTPResult(data: data, response: httpResponse)
-        } catch let error as HTTPClientError {
-            throw error
-        } catch {
-            throw HTTPErrorMapper.map(error, response: nil)
-        }
-    }
-}
-
-struct HTTPResult {
-    let data: Data
-    let response: HTTPURLResponse
-}
-
-protocol HTTPRequest {
-    var host: String { get }
-    var path: String { get }
-    var method: HTTPMethod { get }
-    var headers: [String: String] { get }
-    var body: HTTPBody { get }
-}
-
-extension HTTPRequest {
-    func asURLRequest() throws -> URLRequest {
-        var components = URLComponents()
-        components.scheme = "https"
-        components.host = host
-        components.path = path
-        var request = URLRequest(url: components.url!)
-        request.httpMethod = method.rawValue
-        request.allHTTPHeaderFields = headers
-        request.httpBody = try body.encode()
-        return request
-    }
-}
-
-struct HTTPBody {
-    let defaultHeaders: [String: String]
-    let encode: () throws -> Data
-    
-    static func empty() -> HTTPBody {
-        return HTTPBody(defaultHeaders: [:], encode: { Data() })
-    }
-}
-
-enum HTTPMethod: String {
-    case get = "GET"
-    case post = "POST"
-    case put = "PUT"
-    case delete = "DELETE"
-}
-
-enum HTTPClientError: LocalizedError, Equatable {
-    case invalidURL
-    case networkError(underlying: Error)
-    case serverError(statusCode: Int)
-    case clientError(statusCode: Int)
-    case timeout
-    case noConnection
-    case unauthorized
-    case forbidden
-    case notFound
-    case internalServerError
-    case unknown
-    
-    var errorDescription: String? {
-        switch self {
-        case .invalidURL:
-            return "Invalid URL provided"
-        case .networkError(let underlying):
-            return "Network error: \(underlying.localizedDescription)"
-        case .serverError(let statusCode):
-            return "Server error with status code: \(statusCode)"
-        case .clientError(let statusCode):
-            return "Client error with status code: \(statusCode)"
-        case .timeout:
-            return "Request timed out"
-        case .noConnection:
-            return "No internet connection"
-        case .unauthorized:
-            return "Unauthorized access"
-        case .forbidden:
-            return "Access forbidden"
-        case .notFound:
-            return "Resource not found"
-        case .internalServerError:
-            return "Internal server error"
-        case .unknown:
-            return "Unknown error occurred"
-        }
-    }
-    
-    static func == (lhs: HTTPClientError, rhs: HTTPClientError) -> Bool {
-        return lhs.errorDescription == rhs.errorDescription
-    }
-}
-
-struct HTTPErrorMapper {
-    static func map(_ error: Error, response: URLResponse?) -> HTTPClientError {
-        // Handle URLSession errors
-        if let urlError = error as? URLError {
-            return mapURLError(urlError)
-        }
-        
-        // Handle other errors
-        return .networkError(underlying: error)
-    }
-    
-    private static func mapURLError(_ error: URLError) -> HTTPClientError {
-        switch error.code {
-        case .notConnectedToInternet, .networkConnectionLost:
-            return .noConnection
-        case .timedOut:
-            return .timeout
-        case .badURL, .unsupportedURL:
-            return .invalidURL
-        default:
-            return .networkError(underlying: error)
-        }
-    }
-    
-    static func mapHTTPResponse(_ response: HTTPURLResponse) -> HTTPClientError {
-        switch response.statusCode {
-        case 401:
-            return .unauthorized
-        case 403:
-            return .forbidden
-        case 404:
-            return .notFound
-        case 500...599:
-            return .internalServerError
-        case 400...499:
-            return .clientError(statusCode: response.statusCode)
-        default:
-            return .unknown
-        }
-    }
-}
 
 final class URLSessionHttpClientTests: XCTestCase {
     override func setUp() {
